bot/requirements.txt
discord.py
python-dotenv
psycopg2-binary
asyncpg
